Fix tests after rebase
